feat(quaint): add "provider()" method to "ExternalConnector"
