+++ quaint/src/connector/external.rs
@@ -1,11 +1,65 @@
-use std::sync::Arc;
+use std::{str::FromStr, sync::Arc};
 
 use async_trait::async_trait;
 
 use super::{SqlFamily, TransactionCapable};
 
+#[cfg_attr(target_arch = "wasm32", derive(serde::Deserialize))]
+#[derive(Debug, Eq, PartialEq, Copy, Clone)]
+pub enum AdapterProvider {
+    #[cfg(feature = "mysql")]
+    Mysql,
+    #[cfg(feature = "postgresql")]
+    Postgres,
+    #[cfg(feature = "sqlite")]
+    Sqlite,
+}
+
+impl AdapterProvider {
+    pub fn db_system_name(&self) -> &'static str {
+        match self {
+            #[cfg(feature = "mysql")]
+            Self::Mysql => "mysql",
+            #[cfg(feature = "postgresql")]
+            Self::Postgres => "postgresql",
+            #[cfg(feature = "sqlite")]
+            Self::Sqlite => "sqlite",
+        }
+    }
+}
+
+impl FromStr for AdapterProvider {
+    type Err = String;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            #[cfg(feature = "postgresql")]
+            "postgres" => Ok(Self::Postgres),
+            #[cfg(feature = "mysql")]
+            "mysql" => Ok(Self::Mysql),
+            #[cfg(feature = "sqlite")]
+            "sqlite" => Ok(Self::Sqlite),
+            _ => Err(format!("Unsupported adapter flavour: {:?}", s)),
+        }
+    }
+}
+
+impl From<&AdapterProvider> for SqlFamily {
+    fn from(value: &AdapterProvider) -> Self {
+        match value {
+            #[cfg(feature = "mysql")]
+            AdapterProvider::Mysql => SqlFamily::Mysql,
+            #[cfg(feature = "postgresql")]
+            AdapterProvider::Postgres => SqlFamily::Postgres,
+            #[cfg(feature = "sqlite")]
+            AdapterProvider::Sqlite => SqlFamily::Sqlite,
+        }
+    }
+}
+
 #[derive(Debug, Clone)]
 pub struct ExternalConnectionInfo {
+    // TODO: `sql_family` doesn't exist in TypeScript's `ConnectionInfo` type.
     pub sql_family: SqlFamily,
     pub schema_name: String,
     pub max_bind_values: Option<usize>,
@@ -23,6 +77,7 @@ impl ExternalConnectionInfo {
 
 #[async_trait]
 pub trait ExternalConnector: TransactionCapable {
+    fn provider(&self) -> AdapterProvider;
     async fn get_connection_info(&self) -> crate::Result<ExternalConnectionInfo>;
     async fn execute_script(&self, script: &str) -> crate::Result<()>;
     async fn dispose(&self) -> crate::Result<()>;
